+++ aepp/catalog.py
@@ -170,7 +170,6 @@ class Catalog:
         Returns the last batch from a specific datasetId.
         Arguments:
             dataSetId : OPTIONAL : the datasetId to be retrieved the batch about
-
         """
         path = "/lastBatches"
         params = {}
@@ -247,7 +246,7 @@ class Catalog:
             self.logger.debug(f"Starting getFailedBatchesDF")
         dict_failed = {}
         for batch in res:
-            if res[batch]['relatedObjects'][0]['type'] == "dataSet":
+            if res.get(batch,{}).get('relatedObjects',[{'type':'unknown'}])[0]['type'] == "dataSet":
                 datasetId = res[batch]['relatedObjects'][0]['id']
             dict_failed[batch] = {
                 "timestamp" : res[batch]['created'],