+++ tests/exportDatasetToDatalandingZone_test.py
@@ -1,0 +1,162 @@
+#  Copyright 2023 Adobe. All rights reserved.
+#  This file is licensed to you under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License. You may obtain a copy
+#  of the License at http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software distributed under
+#  the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+#  OF ANY KIND, either express or implied. See the License for the specific language
+#  governing permissions and limitations under the License.
+
+from aepp.exportDatasetToDataLandingZone import ExportDatasetToDataLandingZone
+import unittest
+from unittest.mock import patch, MagicMock, ANY
+from tenacity import RetryError
+
+
+class ExportDatasetToDataLandingZoneTest(unittest.TestCase):
+    flow_response = {
+        "id": "df193495-9837-407f-9d50-e51efc067cb5",
+        "sourceConnectionIds": [
+            "78754315-852c-4c4e-8f1a-d12b83264f3f"
+        ],
+        "state": "disabled",
+        "etag": "test_etag"}
+
+    connection_spec = {
+        "name": "test",
+        "id": "test_connection_id"
+    }
+
+
+    source_connection = {
+        "id": "test_source_connection_id",
+        "params": {
+            "datasets": [{
+                "dataSetId": "test"
+            }]}
+    }
+
+    base_connection = {
+        "id": "test_base_connection_id"
+    }
+
+    target_connection = {
+        "id": "test_target_connection_id"
+    }
+
+    flow_run = {
+        "items": [
+            {
+                "id": "test_id",
+                "metrics": {
+                    "durationSummary": {
+                        "startedAtUTC": 1691184699,
+                        "completedAtUTC": 1691184819
+                    },
+                    "fileSummary": {
+                        "outputFileCount": 24
+                    },
+                    "sizeSummary": {
+                        "outputBytes": 77494320
+                    },
+                    "recordSummary": {
+                        "outputRecordCount": 1065802
+                    }
+                }
+            }
+        ]
+    }
+
+    adhoc_success_response = {
+        "destinations":[
+            {
+                "datasets":[
+                    {
+                        "id":"641ce00b8f31811b98dd3b56",
+                        "statusURL":"https: //platform-stage.adobe.io/data/foundation/flowservice/runs/aa39ad3d-24ba-4579-9cdc-55536c408721",
+                        "flowId":"3eaa2c0b-e24b-46bd-8eee-bbaf9d6cf2cb"
+                    }
+                ]
+            }
+        ]
+    }
+
+
+    adhoc_non_retry_error = {
+        "error_code": "401013",
+        "message": "Oauth token is not valid"
+    }
+
+    adhoc_retry_error = {
+        "message": "Following order ID(s) are not ready for dataset export"
+    }
+
+    config = {
+        "org_id": "3ADF23C463D98F640A494032@AdobeOrg",
+        "client_id": "35e6e4d205274c4ca1418805ac41153b",
+        "tech_id": "test005@techacct.adobe.com",
+        "pathToKey": "/Users/Downloads/config/private.key",
+        "auth_code": "",
+        "secret": "test",
+        "date_limit": 0,
+        "sandbox": "prod",
+        "environment": "stage",
+        "token": "token",
+        "jwtTokenEndpoint": "https://ims-na1.adobelogin.com/ims/exchange/jwt/",
+        "oauthTokenEndpoint": "",
+        "imsEndpoint": "https://ims-na1-stg1.adobelogin.com",
+        "private_key": ""
+    }
+
+    @patch('aepp.utils.Utils.check_if_exists', MagicMock(return_value = "test_dataflow_id"))
+    @patch('aepp.flowservice.FlowService.getFlow', MagicMock(return_value = flow_response))
+    @patch('aepp.flowservice.FlowService.getSourceConnection', MagicMock(return_value = source_connection))
+    @patch('aepp.flowservice.FlowService.getRun', MagicMock(return_value = flow_run))
+    @patch('aepp.destinationinstanceservice.DestinationInstanceService.createAdHocDatasetExport', MagicMock(return_value = adhoc_success_response))
+    @patch("aepp.connector.AdobeRequest", MagicMock())
+    def test_create_dataflow_if_exist(self):
+        export_obj = ExportDatasetToDataLandingZone(config= self.config, header= MagicMock())
+        export_obj.createDataFlowRunIfNotExists(dataset_id="test", compression_type="gzip", data_format="parquet", export_path="test", on_schedule=False, config_path="test", entity_name="test", initial_delay=0)
+
+    @patch('aepp.utils.Utils.check_if_exists', MagicMock(return_value = "test_dataflow_id"))
+    @patch('aepp.flowservice.FlowService.getFlow', MagicMock(return_value = flow_response))
+    @patch('aepp.flowservice.FlowService.getSourceConnection', MagicMock(return_value = source_connection))
+    @patch("aepp.connector.AdobeRequest", MagicMock())
+    def test_create_dataflow_on_schedule(self):
+        export_obj = ExportDatasetToDataLandingZone(config= self.config, header= MagicMock())
+        export_obj.createDataFlowRunIfNotExists(dataset_id="test", compression_type="gzip", data_format="parquet", export_path="test", on_schedule=True, config_path="test",entity_name="test", initial_delay=0)
+
+    @patch('aepp.flowservice.FlowService.createConnection', MagicMock(return_value = base_connection))
+    @patch('aepp.flowservice.FlowService.getConnectionSpecIdFromName', MagicMock(return_value = connection_spec))
+    @patch('aepp.flowservice.FlowService.createSourceConnection', MagicMock(return_value = source_connection))
+    @patch('aepp.flowservice.FlowService.createTargetConnection', MagicMock(return_value = target_connection))
+    @patch('aepp.flowservice.FlowService.createFlow', MagicMock(return_value = flow_response))
+    @patch('aepp.utils.Utils.save_field_in_config', MagicMock())
+    @patch("aepp.connector.AdobeRequest", MagicMock())
+    def test_create_dataflow_if_not_exist(self):
+        export_obj = ExportDatasetToDataLandingZone(config= self.config, header= MagicMock())
+        return_dataflow_id = export_obj.createDataFlow(dataset_id="test", compression_type="gzip", data_format="parquet", export_path="test", on_schedule=False, config_path="test", entity_name="test")
+        assert (return_dataflow_id == self.flow_response["id"])
+
+    @patch('aepp.destinationinstanceservice.DestinationInstanceService.createAdHocDatasetExport', MagicMock(return_value = adhoc_success_response))
+    @patch("aepp.connector.AdobeRequest", MagicMock())
+    def test_retry_on_success_response(self):
+        export_obj = ExportDatasetToDataLandingZone(config= self.config, header= MagicMock())
+        assert(export_obj.retryOnNotReadyException("test", "test", 1, 1) == self.adhoc_success_response)
+
+    @patch('aepp.destinationinstanceservice.DestinationInstanceService.createAdHocDatasetExport', MagicMock(return_value = adhoc_non_retry_error))
+    @patch("aepp.connector.AdobeRequest", MagicMock())
+    def test_no_retry_error(self):
+        export_obj = ExportDatasetToDataLandingZone(config= self.config, header= MagicMock())
+        assert (export_obj.retryOnNotReadyException("test", "test", 1, 1) == self.adhoc_non_retry_error)
+
+    @patch('aepp.destinationinstanceservice.DestinationInstanceService.createAdHocDatasetExport', MagicMock(return_value = adhoc_retry_error))
+    @patch("aepp.connector.AdobeRequest", MagicMock())
+    def test_retry_error(self):
+        export_obj = ExportDatasetToDataLandingZone(config= self.config, header= MagicMock())
+        try:
+            export_obj.retryOnNotReadyException("test", "test", 1, 1)
+            self.fail("expect a retry error")
+        except RetryError:
+            pass